refa: enhance class inheritance
