+++ core/src/main/java/edu/pku/code2graph/model/URI.java
@@ -9,7 +9,7 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 /** Unified Resource Identifier for code elements */
-public class URI extends URILike<Layer>  implements Serializable {
+public class URI extends URIBase<Layer> implements Serializable {
   public URI() {
     this(false, "");
   }
@@ -85,6 +85,16 @@ public class URI extends URILike<Layer>  implements Serializable {
       addLayer("");
     }
     layers.get(2).setIdentifier(identifier);
+  }
+
+  @Override
+  public int hashCode() {
+    return toString().hashCode();
+  }
+
+  @Override
+  public boolean equals(Object obj) {
+    return toString().equals(obj.toString());
   }
 
   private static List<String> pre = Arrays.asList("\\*", "\\(", "\\)", "\\/", "\\[", "\\]");
