+++ xll/src/main/java/edu/pku/code2graph/xll/Scanner.java
@@ -3,16 +3,13 @@ package edu.pku.code2graph.xll;
 import edu.pku.code2graph.model.Layer;
 import edu.pku.code2graph.model.URI;
 import edu.pku.code2graph.model.URITree;
+import edu.pku.code2graph.xll.pattern.URIPattern;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.*;
 
 public class Scanner {
-  private final static Logger logger = LoggerFactory.getLogger(Scanner.class);
-
   /**
    * uri pattern cache
    */
