+++ xll/src/main/java/edu/pku/code2graph/xll/pattern/URIPattern.java
@@ -1,14 +1,15 @@
-package edu.pku.code2graph.xll;
+package edu.pku.code2graph.xll.pattern;
 
 import edu.pku.code2graph.model.Language;
 import edu.pku.code2graph.model.URI;
-import edu.pku.code2graph.model.URILike;
+import edu.pku.code2graph.model.URIBase;
+import edu.pku.code2graph.xll.Capture;
 
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 
-public class URIPattern extends URILike<LayerPattern> {
+public class URIPattern extends URIBase<LayerPattern> {
   public final Set<String> symbols = new HashSet<>();
   public final Set<String> anchors = new HashSet<>();
